Change images, minor adjustments

+++ core/src/com/dcasado/defender/model/components/ShooterComponent.java
@@ -6,6 +6,6 @@ import com.badlogic.ashley.core.Component;
  * Created by David on 23/01/2016.
  */
 public class ShooterComponent implements Component{
-    public int interval;
+    public float interval;
     public float temp;
 }
